Added functionality for messaging to UserBookFragment and MyBookTransactionFragment. Refactor ChatRoom.java to ChatRoomList.java.
