+++ doc/JavaDoc/ca/ualberta/CMPUT3012019T02/alexandria/ChatRoomTests.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.CMPUT3012019T02.alexandria.ChatRoomTests</li>
+<li>ca.ualberta.CMPUT3012019T02.alexandria.ChatRoomListTests</li>
 </ul>
 </li>
 </ul>
