+++ doc/JavaDoc/ca/ualberta/CMPUT3012019T02/alexandria/model/chatroom/ChatRoom.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.CMPUT3012019T02.alexandria.model.chatroom.ChatRoom</li>
+<li>ca.ualberta.CMPUT3012019T02.alexandria.model.chatroom.ChatRoomList</li>
 </ul>
 </li>
 </ul>
@@ -157,7 +157,7 @@ extends java.lang.Object</pre>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code>java.lang.String</code></td>
 <td class="colLast"><code><span class="memberNameLink"><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/model/chatroom/ChatRoom.html#getId--">getId</a></span>()</code>
-<div class="block">gets id of the chatRoom</div>
+<div class="block">gets id of the chatRoomList</div>
 </td>
 </tr>
 <tr id="i2" class="altColor">
@@ -175,7 +175,7 @@ extends java.lang.Object</pre>
 <tr id="i4" class="altColor">
 <td class="colFirst"><code>void</code></td>
 <td class="colLast"><code><span class="memberNameLink"><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/model/chatroom/ChatRoom.html#setId-java.lang.String-">setId</a></span>(java.lang.String&nbsp;id)</code>
-<div class="block">sets id of the chatRoom</div>
+<div class="block">sets id of the chatRoomList</div>
 </td>
 </tr>
 <tr id="i5" class="rowColor">
@@ -293,7 +293,7 @@ extends java.lang.Object</pre>
 <li class="blockList">
 <h4>getId</h4>
 <pre>public&nbsp;java.lang.String&nbsp;getId()</pre>
-<div class="block">gets id of the chatRoom</div>
+<div class="block">gets id of the chatRoomList</div>
 <dl>
 <dt><span class="returnLabel">Returns:</span></dt>
 <dd>String id</dd>
@@ -307,7 +307,7 @@ extends java.lang.Object</pre>
 <li class="blockList">
 <h4>setId</h4>
 <pre>public&nbsp;void&nbsp;setId(java.lang.String&nbsp;id)</pre>
-<div class="block">sets id of the chatRoom</div>
+<div class="block">sets id of the chatRoomList</div>
 <dl>
 <dt><span class="paramLabel">Parameters:</span></dt>
 <dd><code>id</code> - idw</dd>
