+++ doc/JavaDoc/index-files/index-13.html
@@ -168,7 +168,7 @@
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/chatroom/ChatRoom.html#setId-java.lang.String-">setId(String)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.model.chatroom.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/chatroom/ChatRoom.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.model.chatroom">ChatRoom</a></dt>
 <dd>
-<div class="block">sets id of the chatRoom</div>
+<div class="block">sets id of the chatRoomList</div>
 </dd>
 <dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/message/Message.html#setId-java.lang.String-">setId(String)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.model.message.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/message/Message.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.model.message">Message</a></dt>
 <dd>
