+++ doc/JavaDoc/index-files/index-6.html
@@ -133,7 +133,7 @@
 </dd>
 <dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/chatroom/ChatRoom.html#getId--">getId()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.model.chatroom.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/chatroom/ChatRoom.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.model.chatroom">ChatRoom</a></dt>
 <dd>
-<div class="block">gets id of the chatRoom</div>
+<div class="block">gets id of the chatRoomList</div>
 </dd>
 <dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/message/Message.html#getId--">getId()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.model.message.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/message/Message.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.model.message">Message</a></dt>
 <dd>
